fix(Text.stories): adjust the argTypes within the story object
