+++ src/components/data/Text/stories/Text.Dafault.stories.tsx
@@ -1,0 +1,40 @@
+import { Text } from "..";
+
+import { StyledContainer } from "./styles";
+import { parameters, props } from "./props";
+import { ITextProps } from "../interfaces/Text.interface";
+
+console.log(props);
+
+const story = {
+  title: "data/Text/Default",
+  components: [Text],
+  parameters,
+  argTypes: props,
+};
+
+const Default = (args: ITextProps) => {
+  return (
+    <StyledContainer>
+      <Text {...args}>{args.children}</Text>
+    </StyledContainer>
+  );
+};
+
+Default.args = {
+  children: "Title with the Text component",
+  as: "h1",
+  textAlign: "start",
+  margin: "20px 22px 23px 24px",
+  padding: "5px",
+  appearance: "dark",
+  type: "body",
+  size: "large",
+  isDisabled: false,
+  cursorHover: true,
+  ellipsis: false,
+};
+
+export default story;
+
+export { Default };
